Update SendMailHelper.php
It might be worth pointing out in the DocBlock that the EmailList variable can be a string or an array. When I pass the email list as an array to the send method, the IDE shows a warning that there must be a string. Although this method, judging by the code, can also accept an array.