Update AbstractModelRelationHandler.php
Some belongsToMany relationship widgets always submit their content when the form is saved. In cases where such a widget is empty, this can be interpreted as either a complete detach relationship or an empty sheet. In earlier versions, the pivot check was carried out in the detach method, but now there is no such check, and for example, when an empty list of additional categories is saved, an error occurs, or the cache flush does not work because there is no list of detached entities. These changes solve this problem by checking whether bound entities have been deleted or were initially empty.

https://github.com/octobercms/library/blob/1.0/src/Database/Relations/BelongsToMany.php#L166-L168